// generated/api-client/models/feedback-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { FeedbackResponseDtoAttendee } from './feedback-response-dto-attendee';
// May contain unused imports in some cases
// @ts-ignore
import { FeedbackResponseDtoUser } from './feedback-response-dto-user';

/**
 * 
 * @export
 * @interface FeedbackResponseDto
 */
export interface FeedbackResponseDto {
    /**
     * Response identifier
     * @type {string}
     * @memberof FeedbackResponseDto
     */
    'id': string;
    /**
     * User responses to form fields
     * @type {object}
     * @memberof FeedbackResponseDto
     */
    'responses': object;
    /**
     * When the response was submitted
     * @type {string}
     * @memberof FeedbackResponseDto
     */
    'submittedAt': string;
    /**
     * 
     * @type {FeedbackResponseDtoUser}
     * @memberof FeedbackResponseDto
     */
    'user': FeedbackResponseDtoUser;
    /**
     * 
     * @type {FeedbackResponseDtoAttendee}
     * @memberof FeedbackResponseDto
     */
    'attendee': FeedbackResponseDtoAttendee;
}